refactor: 전역 scss에서 'pretendard' -> 'Pretendard' 이름 변경
