+++ src/styles/_reset.scss
@@ -2,6 +2,7 @@
 *::before,
 *::after {
   box-sizing: border-box;
+  font-family: $font-family-primary;
 }
 html,
 body,
