+++ src/styles/_token.scss
@@ -52,7 +52,7 @@ $font-2xl: 1.5rem; /* 24px */
 $font-3xl: 1.75rem; /* 28px */
 
 /* Font Family */
-$font-family-primary: 'pretendard';
+$font-family-primary: 'Pretendard';
 
 /* Spacing */
 $size-xs: 0.3125rem; /* 5px */
